% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{badmf.sims.2ball}
\alias{badmf.sims.2ball}
\title{Sample from Unit 2-Ball}
\usage{
badmf.sims.2ball(n, d, r = 1, cov.scale = 0)
}
\arguments{
\item{n}{the number of samples.}

\item{d}{the number of dimensions.}

\item{r}{the radius of the 2-ball. Defaults to \code{1}.}

\item{cov.scale}{if desired, sample from 2-ball with error sigma. Defaults to \code{NaN},
which has no noise.}
}
\description{
Sample from the 2-ball in d-dimensions.
}
\examples{
library(badmf)
# sample 100 points from 3-d 2-ball with radius 2
X <- badmf.sims.rball(100, 3, 2)
}
\author{
Eric Bridgeford
}
