% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{badmf.sims.cross}
\alias{badmf.sims.cross}
\title{Cross Simulation}
\usage{
badmf.sims.cross(n, d, K = 2, signal.scale = 10, non.scale = 1,
  rotate = FALSE, class.equal = TRUE, ind = FALSE)
}
\arguments{
\item{n}{the number of samples.}

\item{d}{the number of dimensions.}

\item{K}{the number of classes in the dataset.}

\item{signal.scale}{the scaling for the signal dimension. Defaults to \code{10}.}

\item{non.scale}{the scaling for the non-signal dimensions. Defaults to \code{1}.}

\item{class.equal}{whether the number of samples/class should be equal, with each
class having a prior of 1/K, or inequal, in which each class obtains a prior
of k/sum(K) for k=1:K. Defaults to \code{TRUE}.}

\item{ind}{whether to sample x and y independently. Defaults to \code{FALSE}.}
}
\description{
A function to simulate data with a crossed layout.
}
\examples{
library(badmf)
sim <- badmf.sims.cross(100, 3)
X <- sim$X; Y <- sim$Y
}
\author{
Eric Bridgeford
}
