% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{badmf.sims.random_rotate}
\alias{badmf.sims.random_rotate}
\title{Random Rotation}
\usage{
badmf.sims.random_rotate(mus, Sigmas, Q = NULL)
}
\arguments{
\item{mus}{means per class.}

\item{Sigmas}{covariances per class.}

\item{Q}{rotation to use, if any}
}
\description{
A helper function for applying a random rotation to gaussian parameter set.
}
\author{
Eric Bridgeford
}
