% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{build.tree}
\alias{build.tree}
\title{Recursive Approach for Building Decision Tree}
\usage{
build.tree(split, d, alpha, depth.max, size, depth, debug = FALSE)
}
\arguments{
\item{split}{a particular split node.}

\item{d}{the number of features to subsample.}

\item{alpha}{the sampling distribution for the features. A \code{[p]} vector. If \code{NULL}, uses a uniform.}

\item{depth.max}{the max tree depth.}

\item{size}{the minimum number of elements at a particular.}

\item{depth}{the current depth of the tree.}

\item{debug}{a boolean indicating whether to save the predictors and responses
that are categorized into leaf nodes.}
}
\value{
a layer of a decision tree.
}
\description{
Recursive Approach for Building Decision Tree
}
\author{
Eric Bridgeford
}
