% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{count.features.dec.tree.class}
\alias{count.features.dec.tree.class}
\title{Count Features in a Decision Tree}
\usage{
\method{count.features}{dec.tree.class}(object, ...)
}
\arguments{
\item{object}{an object of class \code{dec.tree.class}.}

\item{...}{trailing args.}
}
\value{
a vector containing the number of times each feature
is used at a split node.
}
\description{
A function to count the feature utilization in a Decision Tree.
}
\author{
Eric Bridgeford
}
