% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{count.features.leaf.node}
\alias{count.features.leaf.node}
\title{Count Features Leaf}
\usage{
\method{count.features}{leaf.node}(object, ...)
}
\arguments{
\item{object}{an object of class \code{leaf.node}.}

\item{...}{trailing args.}
}
\value{
a vector containing the number of times each feature is used
at a split node.
}
\description{
A default to return NULL and stop recursion when counting features
in a stump once we reach a leaf node.
}
\author{
Eric Bridgeford
}
