% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{count.features}
\alias{count.features}
\title{Counting Feature Utilization}
\usage{
count.features(object, ...)
}
\arguments{
\item{object}{a model object for which feature counting
is desired.}

\item{...}{additional arguments impacting feature counting.}
}
\value{
the counts of features used in the tree or forest.
}
\description{
A generic method for counting features in a
Tree or Forest.
}
\author{
Eric Bridgeford
}
