% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{count.features.rf.class}
\alias{count.features.rf.class}
\title{Count Feature Utilization}
\usage{
\method{count.features}{rf.class}(object, ...)
}
\arguments{
\item{object}{a fit random forest of class \code{rf.class}.}

\item{...}{trailing args.}
}
\value{
the counts of all features in the forest.
}
\description{
A function to count the number of times a feature is used
in a forest at split nodes.
}
\author{
Eric Bridgeford
}
