% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{count.features.split.node}
\alias{count.features.split.node}
\title{Count Features Stump}
\usage{
\method{count.features}{split.node}(object, ...)
}
\arguments{
\item{object}{an object of class \code{split.node}.}

\item{...}{trailing args.}
}
\value{
a vector containing the number of times each feature is used
at a split node.
}
\description{
A function to count the Usage of Features At or Below a Split Node (a Stump)
}
\author{
Eric Bridgeford
}
