% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{create.split}
\alias{create.split}
\title{Create a new split of the data.}
\usage{
create.split(X, Y, i, t)
}
\arguments{
\item{X}{the predictors.}

\item{Y}{the responses.}

\item{i}{the feature index.}

\item{t}{the threshold for the feature.}
}
\value{
the split, as a list.
}
\description{
Create a new split of the data.
}
\author{
Eric Bridgeford
}
