% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{dec_tree.fit}
\alias{dec_tree.fit}
\title{Bayesian Decision Tree Fit}
\usage{
dec_tree.fit(formuler, data = NULL, family = "dirichlet",
  prior = NULL)
}
\arguments{
\item{formuler}{ravioli ravioli give me the formuoli.}

\item{data}{the data associated with the formuler.}

\item{family}{the family to use for the feature prior. If NULL, does not construct a prior.}
}
\value{
A trained decision tree.
}
\description{
Fit a Bayesian Decision Tree.
}
\author{
Eric Bridgeford
}
