% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{impurity.idx}
\alias{impurity.idx}
\title{A function for computing the impurity of a particular split.}
\usage{
impurity.idx(groups)
}
\arguments{
\item{groups}{a 2 element list, where the first element are the samples to the left
and the second element the samples to the right of a split.}
}
\value{
the gini impurity index.
}
\description{
A function for computing the impurity of a particular split.
}
\author{
Eric Bridgeford
}
