% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{leaf.node}
\alias{leaf.node}
\title{Return Most Probable Class at a Leaf Node}
\usage{
leaf.node(Y)
}
\arguments{
\item{Y}{the responses at a leaf node.}
}
\value{
the most probable response at the leaf.
}
\description{
Return Most Probable Class at a Leaf Node
}
