% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dec.tree.class}
\alias{predict.dec.tree.class}
\title{Decision Tree Prediction}
\usage{
\method{predict}{dec.tree.class}(object, X, ...)
}
\arguments{
\item{object}{a fit decision tree.}

\item{X}{an \code{[n, p]} predictor to predict the class of. Must be specified in an appropriate
2d array, even if you only want a single sample predicted.}
}
\value{
an \code{[n]} array of predictions.
}
\description{
A method for Decision Tree prediction. Predicts by pushing samples down
to a leaf node based on the split node criterion.
}
\author{
Eric Bridgeford
}
