% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dec.tree}
\alias{predict.dec.tree}
\title{Decision Tree Prediction}
\usage{
\method{predict}{dec.tree}(object, X, ...)
}
\arguments{
\item{object}{a fit decision tree.}

\item{X}{an [n, p] predictor to predict the class of. Must be specified in an appropriate
2d array, even if you only want a single sample predicted.}
}
\value{
an [n] array of predictions.
}
\description{
A helper method for Decision Tree prediction.
}
\author{
Eric Bridgeford
}
