% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.leaf.node}
\alias{predict.leaf.node}
\title{A function for leaf node prediction}
\usage{
\method{predict}{leaf.node}(object, ...)
}
\arguments{
\item{object}{a leaf.node}
}
\value{
the vote for the leaf.
}
\description{
A function for leaf node prediction
}
