% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rf.class}
\alias{predict.rf.class}
\title{Random Forest Prediction}
\usage{
\method{predict}{rf.class}(object, X, ...)
}
\arguments{
\item{object}{a fit object of class rf.class.}

\item{X}{an \code{[n, p]} predictor to predict the class of. Must be specified in an appropriate
2d array, even if you only want a single sample predicted.}
}
\value{
an \code{[n]} array of predictions.
}
\description{
A method for Random Forest prediction. Predicts by, for each tree in the forest,
doing a standard decision tree prediction. Then, the most likely vote across the trees
is used for predictions.
}
\details{
If the two or more classes are tied for the maximum number of votes across the trees,
the predicted class is chosen amongst the classes tied for the maximal number of votes
at random.
}
\author{
Eric Bridgeford
}
