% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf.fit}
\alias{rf.fit}
\title{Decision Tree Frontend Interface}
\usage{
rf.fit(formula, data, family = "dirichlet", method = "forest.fit",
  prior = NULL)
}
\arguments{
\item{data}{the data associated with the formula.}

\item{family}{the family to use for the feature prior. If NULL, does not construct a prior.}

\item{form}{a valid formula object.}
}
\value{
A trained decision tree.
}
\description{
Decision Tree Frontend Interface
}
\author{
Eric Bridgeford
}
