% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf.input.validator}
\alias{rf.input.validator}
\title{Input Validator}
\usage{
rf.input.validator(d, p, ntrees, bagg, depth.max, depth, size)
}
\arguments{
\item{d}{the number of features to subsample at each node. Defaults to \code{sqrt(p)}.}

\item{p}{the number of features total.}

\item{ntrees}{the number of trees to construct. Defaults to \code{10}.}

\item{bagg}{the relative size of the subsamples for the training set. Defaults to \code{0.632}. A numeric s.t.
\code{0 < bagg <= 1}. Each subsample will be \code{bagg*n} elements.}

\item{depth.max}{the maximum allowed tree depth.}

\item{size}{the minimum allowed number of samples for an individual node.}
}
\value{
nothing; throws an error if the input is invalid.
}
\description{
Input Validator
}
\author{
Eric Bridgeford
}
