% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{tree.class.fit}
\alias{tree.class.fit}
\title{Fit a decision tree classifier.}
\usage{
tree.class.fit(X, Y, d = NULL, alpha = NULL, depth.max = 1,
  size = 1, debug = FALSE)
}
\arguments{
\item{X}{the predictors. A `[n, p]` matrix.}

\item{Y}{the responses. A `[n]` vector or, optionally, a factor.}

\item{d}{the number of features to subsample at each node. Defaults to `NULL`, which tries every feature.}

\item{alpha}{the prior parameters for the feature probabilities. A `[p]` vector. If `NULL`, samples uniformly.}

\item{depth.max}{the maximum allowed tree depth. Defaults to `1`.}

\item{size}{the minimum allowed number of samples for an individual node. Defaults to `1`.}

\item{debug}{whether to save the predictors and responses that are categorized. Defaults to `FALSE`.
into a particular leaf node.}
}
\value{
an object of class `dec.tree` containing the following:
\item{`tree`}{the decision tree.}
\item{`X`}{The training predictors.}
\item{`Y`}{the training responses.}
\item{`d`}{d the number of features subsampled at each node.}
\item{`alpha`}{the sampling distribution for the features. A `[p]` vector.}
\item{`depth.max`}{the maximum allowed tree depth.}
\item{`size`}{the maximum allowed tree depth.}
\item{`debug`}{whether to save the predictors and responses that are categorized.}
}
\description{
Fit a decision tree classifier.
}
\author{
Eric Bridgeford
}
