% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{tree.fit}
\alias{tree.fit}
\title{Fit a tree.}
\usage{
tree.fit(X, Y, d, depth.max = 1, size = 1, debug = FALSE)
}
\arguments{
\item{X}{the predictors.}

\item{Y}{the responses.}

\item{d}{the number of features to subsample.}

\item{size}{the minimum allowed number of samples for an individual node.}

\item{debug}{whether to save the predictors and responses that are categorized
into a particular leaf node.}

\item{depth}{the maximum allowed tree depth.}
}
\value{
a fit decision tree.
}
\description{
Fit a tree.
}
\author{
Eric Bridgeford
}
