% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badmf.R
\name{badmf.fit}
\alias{badmf.fit}
\title{Bayesian Decision-Making Forest (BaD-MF)}
\usage{
badmf.fit(formuler, data = NULL, d = NULL, alpha = NULL,
  ntrees = 10L, bagg = 0.632, method = "classification",
  depth.max = 1L, size = 1L, debug = FALSE, mc.cores = 1L,
  train.params = NULL, ...)
}
\arguments{
\item{formuler}{ravioli ravioli give me the formuoli.}

\item{data}{the data associated with the formuler. Note: if you want an intercept, you must
add it ahead of time.}

\item{d}{the number of features to subsample at a split node. Defaults to \code{sqrt(nsamples)}.}

\item{alpha}{the feature sampling prior. Should be a \code{[p]} vector, where \code{p} is the number of predictors.
Corresponds to alpha for a Dirichlet distribution. If \code{NULL}, samples uniformly for the initial
training iteration.}

\item{ntrees}{the number of trees to construct. Defaults to \code{10L}.}

\item{bagg}{the relative size of the subsamples for the training set. A numeric s.t.
\code{0 < bagg <= 1}. Each subsample will be \code{bagg*n} elements. Defaults to \code{0.632}.}

\item{depth.max}{the maximum allowed tree depth. Defaults to \code{5L}.}

\item{size}{the minimum allowed number of samples for an individual node. Defaults to \code{1L}.}

\item{debug}{whether to save the predictors and responses that are categorized. Defaults to \code{FALSE}.}

\item{mc.cores}{the number of cores to use. Should be \code{0 < mc.cores <= parallel::detectCores()}.
Defaults to \code{1L}.}

\item{train.params}{if you wish to provide specialized parameters for training, a named list containing the following named elements:
\itemize{
\item{\code{d} the number of features to subsample at a split node.}
\item{\code{ntrees} the number of trees to construct.}
\item{\code{bagg} the relative size of the subsamples from the training set.}
\item{\code{depth.max} the maximum allowed tree depth.}
\item{\code{size} the minimum allowed number of samples for an individual node.}
}
Any unset parameters will default to the values provided above (or the corresponding defaults if unprovided).}

\item{...}{trailing arguments.}
}
\value{
an object of class \code{rf.class} containing the following:
\item{\code{forest}}{A list a decision trees.}
\item{\code{method}}{the method used to fit the forest.}
}
\description{
Fit a Bayesian Decision-Making Forest with a \code{stats}-like formula frontend interface.
}
\author{
Eric Bridgeford
}
