% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{dec_tree.fit}
\alias{dec_tree.fit}
\title{Decision Tree}
\usage{
dec_tree.fit(formula, data, family = "dirichlet", method = "tree.fit",
  prior = NULL)
}
\arguments{
\item{data}{the data associated with the formula.}

\item{form}{a valid formula object.}
}
\value{
A trained decision tree.
}
\description{
Decision Tree
}
\author{
Eric Bridgeford
}
