% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tree.R
\name{get.split}
\alias{get.split}
\title{Find Best split in the data.}
\usage{
get.split(X, Y, d)
}
\arguments{
\item{X}{the predictors.}

\item{Y}{the responses.}

\item{d}{the number of features to subsample. Should be an integer 0 < d <= p.}
}
\value{
the best split.
}
\description{
Find Best split in the data.
}
