% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf.fit}
\alias{rf.fit}
\title{Random Forest (RF)}
\usage{
rf.fit(formuler, data = NULL, d = NULL, alpha = NULL, ntrees = 10L,
  bagg = 0.632, method = "classification", depth.max = 1L,
  size = 1L, debug = FALSE, no.cores = 1L)
}
\arguments{
\item{formuler}{ravioli ravioli give me the formuoli.}

\item{data}{the data associated with the formuler. Note: if you want an intercept, you must
add it ahead of time.}

\item{d}{the number of features to subsample at a split node. Defaults to sqrt(nsamples).}

\item{alpha}{the feature sampling prior. Should be a `[p]` vector, where `p` is the number of predictors.
Corresponds to alpha for a Dirichlet distribution.   If `NULL`, samples uniformly.}

\item{ntrees}{the number of trees to construct. Defaults to 10.}

\item{bagg}{the relative size of the subsamples for the training set. Defaults to 0.632. A numeric s.t.
`0 < bagg <= 1`. Each subsample will be `bagg*nsamples`` elements.}

\item{method}{whether you want "classification" or "regression".}

\item{depth.max}{the maximum allowed tree depth.}

\item{size}{the minimum allowed number of samples for an individual node.}

\item{debug}{whether to save the predictors and responses that are categorized}

\item{no.cores}{the number of cores to use. Should be `0 < no.cores <= parallel::detectCores()`.
Any unset parameters will default to the values provided above (or the corresponding defaults if unprovided).}
}
\value{
an object of class `rf` containing the following:
\item{`forest`}{A list a decision trees.}
\item{`method`}{the method used to fit the forest.}
}
\description{
Fit a Random Forest with a `stats`-like formula frontend interface.
}
\author{
Eric Bridgeford
}
