% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf.input.validator}
\alias{rf.input.validator}
\title{Input Validator}
\usage{
rf.input.validator(d, ntrees, bagg, depth.max, depth, size)
}
\arguments{
\item{d}{the number of features to subsample at each node. Defaults to `sqrt(p)`.}

\item{ntrees}{the number of trees to construct. Defaults to 10.}

\item{bagg}{the relative size of the subsamples for the training set. Defaults to 0.632. A numeric s.t.
`0 < bagg <= 1`. Each subsample will be `bagg*n`` elements.}

\item{depth.max}{the maximum allowed tree depth.}

\item{size}{the minimum allowed number of samples for an individual node.}
}
\description{
Input Validator
}
